requirements.txt
# --- core numerics -----------------------------------------------------------
numpy>=1.24
scipy>=1.11
numba>=0.58

# --- machine-learning stack (pure-JAX; no TF) -------------------------------
jax>=0.4          # CPU & CUDA wheels from https://github.com/google/jax
jaxlib>=0.4
dm-haiku>=0.0.11  # lightweight neural-net wrapper
optax>=0.1.7      # optimizers & schedules for JAX

# --- plotting / diagnostics --------------------------------------------------
matplotlib>=3.8

# --- utilities ---------------------------------------------------------------
pandas>=2.2
typing_extensions>=4.11
tqdm>=4.66
